Moved models to their respective app, working code
