+++ userauth/models.py
@@ -1,0 +1,12 @@
+from django.contrib.auth.models import AbstractUser
+
+# Create your models here.
+
+
+class CustomUser(AbstractUser):
+    pass
+
+    # add additional fields in here
+
+    def __str__(self):
+        return self.username
