+++ woolnews_app/forms.py
@@ -1,21 +1,3 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm, UserChangeForm
-from .models import CustomUser
 
-
-class CustomUserCreationForm(UserCreationForm):
-    class Meta:
-        model = CustomUser
-        fields = (
-            'username',
-            'email'
-        )
-
-
-class CustomUserChangeForm(UserChangeForm):
-    class Meta:
-        model = CustomUser
-        fields = (
-            'username',
-            'email'
-        )
+# TODO: Post Form, Comment Form.
