+++ woolnews_app/views.py
@@ -1,12 +1,4 @@
 from django.shortcuts import render
-from django.urls import reverse_lazy
-from django.views.generic.edit import CreateView
-
-from .forms import CustomUserCreationForm
 
 
-# SignUp View
-class SignUpView(CreateView):
-    form_class = CustomUserCreationForm
-    success_url = reverse_lazy('login')
-    template_name = 'registration/signup.html'
+# TODO: Post View
