+++ woolnews_config/urls.py
@@ -20,7 +20,7 @@ from django.views.generic.base import TemplateView
 urlpatterns = [
     path('', TemplateView.as_view(template_name='home.html'), name='home'),
     path('admin/', admin.site.urls),
-    path('accounts/', include('woolnews_app.urls')),
+    path('accounts/', include('userauth.urls')),
     path('accounts/', include('django.contrib.auth.urls')),
 
 ]
